requirements.txt
asgiref==3.7.2
Django==5.0.2
django-hashid-field==3.4.0
django-widget-tweaks==1.5.0
hashids==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
